pyproject.toml
[project]
name = "inoue"
version = "0.1.0"
authors = [{ name = "karin0", email = "karin0@gmx.com" }]
requires-python = ">=3.12"
dependencies = [
    "lark>=1.3.1",
    "simpleeval>=1.0.3",
    "python-telegram-bot>=22",
    "prettytable>=3.17.0",
    "yt-dlp>=2026.3.17",
    "pathvalidate>=3.3.1",
    "pillow>=10.0.0",
]
optional-dependencies = { regex = ["regex>=2026.4.4"] }
[dependency-groups]
dev = ["black>=26.3.1"]
